backend/requirements.txt
dspy-ai
fastapi
uvicorn[standard]
python-multipart
pypdf
python-docx
websockets
python-dotenv
